+++ example/lib/models/table_query.dart
@@ -1,7 +1,0 @@
-class TableQuery {
-  final String params;
-  final DateTime start;
-  final DateTime end;
-
-  TableQuery({this.params, this.start, this.end});
-}