+++ lib/simple_timetable.dart
@@ -1,6 +1,6 @@
 library simple_timetable;
 
 export 'src/event.dart';
-export 'src/timetable_view.dart';
-export 'src/timetable_helper.dart';
 export 'src/timetable_direction.dart';
+export 'src/timetable_helper.dart';
+export 'src/timetable_view.dart';