+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.0.11
+
+## Feat
+
+- Add lint
+- Add customazing controls button prev/next
+- Rebuild table on change visibleRange
+- Update package dart_date
+
 # 0.0.10
 
 ## Feat